chore: update dependencies to latest versions
Update several workspace dependencies from version 0.6.6 to 0.6.7. This
includes packages such as rspack_binding_api, rspack_core, and others.
Additionally, update the @rspack/core catalog version in the pnpm
workspace to 1.6.7. The support for the new is-terminal and supports-color
dependencies is also added. This update helps to use the latest features
and fixes.